앱 별로 만들었던 models들을 api 디렉토리에 합쳐놓았습니다
