+++ api/admin.py
@@ -1,16 +1,3 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.contrib.auth.models import User
-from .models import Profile
-
-class ProfileInline(admin.StackedInline):
-    model = Profile
-    can_delete = False
-    verbose_name_plural = 'profile'
-
-
-class UserAdmin(BaseUserAdmin):
-    inlines = (ProfileInline, )
-
-
-admin.site.register(Profile)
